change to search activation phase 2
