+++ packages/compass-telemetry/src/growth-experiments.ts
@@ -1,8 +1,7 @@
 export const ExperimentTestNames = {
   mockDataGenerator: 'MOCK_DATA_GENERATOR_ITERATION_20260507',
   searchActivationProgramP1: 'SEARCH_ACTIVATION_PROGRAM_P1_20260427',
-  searchContextualAiAssistantEntry:
-    'SEARCH_CONTEXTUAL_AI_ASSISTANT_ENTRY_20260608',
+  searchActivationProgramP2: 'SEARCH_ACTIVATION_PROGRAM_P2_20260609',
 } as const;
 
 export type ExperimentTestName =
@@ -13,10 +12,8 @@ export const ExperimentTestGroups = {
   mockDataGeneratorControl: 'mockDataGeneratorControl',
   searchActivationProgramP1Variant: 'searchActivationProgramP1Variant',
   searchActivationProgramP1Control: 'searchActivationProgramP1Control',
-  searchContextualAiAssistantEntryVariant:
-    'searchContextualAiAssistantEntryVariant',
-  searchContextualAiAssistantEntryControl:
-    'searchContextualAiAssistantEntryControl',
+  searchActivationProgramP2Variant: 'searchActivationProgramP2Variant',
+  searchActivationProgramP2Control: 'searchActivationProgramP2Control',
 } as const;
 
 export type ExperimentTestGroup =
