+++ packages/compass-telemetry/src/provider.tsx
@@ -177,5 +177,5 @@ export { useAssignment, useTrackInSample } from './experimentation-provider';
 // @experiment Search Activation Program P1  | Jira Epic: CLOUDP-308952
 export { useSearchActivationProgramP1 } from './search-activation-program-p1';
 
-// @experiment Search Contextual AI Assistant Entry | Jira Epic: CLOUDP-411692
-export { useSearchContextualAiAssistantEntry } from './search-contextual-ai-assistant-entry';
+// @experiment Search Activation Program P2 | Jira Epic: CLOUDP-331931
+export { useSearchActivationProgramP2 } from './search-activation-program-p2';
